# server/go/model_product.go
/*
 * Jumia
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Product struct {

	// The product sku
	Sku string `json:"sku,omitempty"`

	// The product name
	Name string `json:"name,omitempty"`

	// The product country
	Country string `json:"country,omitempty"`
}